chore: move ETT functionality out of permanent_config
