+++ lib/screenplay/emergency_takeover_tool/config_updater.ex
@@ -1,14 +1,103 @@
-defmodule Screenplay.EmergencyTakeoverTool.ScreensConfig do
+defmodule Screenplay.EmergencyTakeoverTool.ConfigUpdater do
   @moduledoc """
   Module responsible for building the EmergencyTakeover struct used in Screens configurations
   of active Emergency Takeover Alerts.
   """
+  require Logger
 
   alias Screenplay.EmergencyTakeoverTool.CannedMessages
   alias Screenplay.EmergencyTakeoverTool.EmergencyTakeover, as: EmergencyTakeoverContext
-  alias ScreensConfig.{EmergencyMessagingLocation, EmergencyTakeover, Screen}
+  alias Screenplay.ScreensConfig.Fetch, as: PublishedScreensFetch
+  alias ScreensConfig.{Config, EmergencyMessagingLocation, EmergencyTakeover, Screen}
 
   @image_store Application.compile_env!(:screenplay, :image_store_module)
+
+  def add_emergency_takeover_configs(alert_id, showtime_screen_ids, message) do
+    with {published_config, _published_version_id} <- get_current_published_config(),
+         {:ok, published_config_deserialized} <- Jason.decode(published_config) do
+      %Config{screens: published_screens, devops: devops} =
+        published_config_deserialized |> Config.from_json()
+
+      updated_screens =
+        update_screens_with_emergency_takeover(
+          published_screens,
+          showtime_screen_ids,
+          alert_id,
+          message
+        )
+
+      %Config{screens: updated_screens, devops: devops}
+      |> publish_new_config()
+    else
+      _error ->
+        {:error, "Could not fetch published screens config"}
+    end
+  end
+
+  defp update_screens_with_emergency_takeover(screens, screen_ids, alert_id, message) do
+    for {id, screen} <- screens,
+        into: %{} do
+      if id in screen_ids do
+        case screen do
+          %Screen{app_params: %{emergency_messaging_location: eml}} when not is_nil(eml) ->
+            emergency_takeover = build_emergency_takeover(message, alert_id, screen.app_id, eml)
+
+            {id,
+             put_in(
+               screen,
+               [Access.key!(:app_params), Access.key!(:emergency_takeover)],
+               emergency_takeover
+             )}
+
+          _ ->
+            Logger.error("Tried to takeover #{id} without an emergency_messaging_location")
+
+            {id, screen}
+        end
+      else
+        {id, screen}
+      end
+    end
+  end
+
+  def clear_emergency_takeover_configs(showtime_screen_ids) do
+    with {published_config, _published_version_id} <- get_current_published_config(),
+         {:ok, published_config_deserialized} <- Jason.decode(published_config) do
+      %Config{screens: published_screens, devops: devops} =
+        published_config_deserialized |> Config.from_json()
+
+      updated_screens = clear_screens_emergency_takeover(published_screens, showtime_screen_ids)
+
+      %Config{screens: updated_screens, devops: devops}
+      |> publish_new_config()
+    else
+      _error ->
+        {:error, "Could not fetch published screens config"}
+    end
+  end
+
+  defp clear_screens_emergency_takeover(screens, screen_ids) do
+    for {id, screen} <- screens, into: %{} do
+      if id in screen_ids do
+        {id, put_in(screen, [Access.key!(:app_params), Access.key!(:emergency_takeover)], nil)}
+      else
+        {id, screen}
+      end
+    end
+  end
+
+  defp get_current_published_config do
+    case PublishedScreensFetch.fetch_config() do
+      {:ok, config, version_id} -> {config, version_id}
+      error -> error
+    end
+  end
+
+  def publish_new_config(new_config) do
+    with(:ok <- PublishedScreensFetch.put_config(new_config)) do
+      PublishedScreensFetch.commit()
+    end
+  end
 
   @spec build_emergency_takeover(
           EmergencyTakeoverContext.message(),
@@ -16,7 +105,7 @@ defmodule Screenplay.EmergencyTakeoverTool.ScreensConfig do
           Screen.app_id(),
           EmergencyMessagingLocation.t()
         ) :: EmergencyTakeover.t()
-  def build_emergency_takeover(message, alert_id, app_id, eml) do
+  defp build_emergency_takeover(message, alert_id, app_id, eml) do
     %EmergencyTakeover{
       audio_asset_path: audio_path(message, eml),
       text_for_audio: text_for_audio(message, eml),
